Revert "ComponentRegistry Abstraction (#8193)"
This reverts commit 92bb3528b6053fef700e0073eebde297946ea9d9.
