+++ lib/streamlit/runtime/runtime.py
@@ -22,8 +22,6 @@ from enum import Enum
 from typing import TYPE_CHECKING, Awaitable, Final, NamedTuple
 
 from streamlit import config
-from streamlit.components.lib.local_component_registry import LocalComponentRegistry
-from streamlit.components.types.base_component_registry import BaseComponentRegistry
 from streamlit.logger import get_logger
 from streamlit.proto.BackMsg_pb2 import BackMsg
 from streamlit.proto.ForwardMsg_pb2 import ForwardMsg
@@ -96,11 +94,6 @@ class RuntimeConfig:
     # The cache storage backend for Streamlit's st.cache_data.
     cache_storage_manager: CacheStorageManager = field(
         default_factory=LocalDiskCacheStorageManager
-    )
-
-    # The ComponentRegistry instance to use.
-    component_registry: BaseComponentRegistry = field(
-        default_factory=LocalComponentRegistry
     )
 
     # The SessionManager class to be used.
@@ -202,7 +195,6 @@ class Runtime:
         self._state = RuntimeState.INITIAL
 
         # Initialize managers
-        self._component_registry = config.component_registry
         self._message_cache = ForwardMsgCache()
         self._uploaded_file_mgr = config.uploaded_file_manager
         self._media_file_mgr = MediaFileManager(storage=config.media_file_storage)
@@ -227,10 +219,6 @@ class Runtime:
     @property
     def state(self) -> RuntimeState:
         return self._state
-
-    @property
-    def component_registry(self) -> BaseComponentRegistry:
-        return self._component_registry
 
     @property
     def message_cache(self) -> ForwardMsgCache:
