+++ lib/tests/streamlit/runtime/runtime_test.py
@@ -22,7 +22,6 @@ from unittest.mock import ANY, MagicMock, call, patch
 
 import pytest
 
-from streamlit.components.lib.local_component_registry import LocalComponentRegistry
 from streamlit.proto.ForwardMsg_pb2 import ForwardMsg
 from streamlit.runtime import (
     Runtime,
@@ -640,7 +639,6 @@ class ScriptCheckTest(RuntimeTestCase):
         config = RuntimeConfig(
             script_path=self._path,
             command_line=None,
-            component_registry=LocalComponentRegistry(),
             media_file_storage=MemoryMediaFileStorage("/mock/media"),
             uploaded_file_manager=MemoryUploadedFileManager("/mock/upload"),
             session_manager_class=MagicMock,
